src/requirements.txt
Flask
requests
markdown2
wikipedia
google-generativeai
gunicorn